Add module build from selection
