+++ nodemapper/src/gui/Builder/components/Flow/Utils.ts
@@ -1,0 +1,50 @@
+import { Node } from 'NodeMap/scene/Flow'; // Custom Node definition
+
+export const getNodeById = (id: string, nodes: Node[]): Node | null => {
+  for (const node of nodes) {
+    if (node.id === id) {
+      return node;
+    }
+  }
+  return null;
+};
+
+export const getNodeName = (node: Node): string => {
+  return node.data.config.name;
+};
+
+export const getNodeByName = (name: string, nodes: Node[]): Node | null => {
+  for (const node of nodes) {
+    if (node.data.config.name === name) {
+      return node;
+    }
+  }
+  return null;
+};
+
+export const setNodeName = (nodes: Node[], id: string, name: string): Node[] =>
+  nodes.map((node) => {
+    if (node.id === id) {
+      const newnode = JSON.parse(JSON.stringify(node));
+      newnode.data.config.name = name;
+      return newnode;
+    } else {
+      return node;
+    }
+  });
+
+export const setNodeWorkflow = (
+  nodes: Node[],
+  id: string,
+  workflow: Record<string, unknown>,
+): Node[] => {
+  return nodes.map((node) => {
+    if (node.id === id) {
+      const newnode = JSON.parse(JSON.stringify(node));
+      newnode.data.config.config = workflow;
+      return newnode;
+    } else {
+      return node;
+    }
+  });
+};
