+++ nodemapper/src/gui/Builder/components/Flow/index.tsx
@@ -1,0 +1,5 @@
+import Flow from './Flow';
+export { Edge, Node } from 'NodeMap/scene/Flow';
+export { getNodeById, getNodeByName, getNodeName, setNodeName, setNodeWorkflow } from './Utils';
+
+export default Flow;
