+++ nodemapper/src/gui/Builder/components/Logger/Logger.tsx
