+++ nodemapper/src/gui/Builder/components/Logger/index.tsx
@@ -1,0 +1,3 @@
+import Logger from './Logger';
+
+export default Logger;
