+++ nodemapper/src/gui/Builder/components/NodeInfo/NodeInfo.tsx
@@ -1,8 +1,8 @@
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import React from 'react';
+import * as styles from './../Flow/styles.module.css';
 import ResizeHandle from './../ResizeHandle';
-import * as styles from './../styles.module.css';
 import ConfigSelect from './ConfigSelect';
 import HighlightedJSON from './HighlightedJSON';
