+++ nodemapper/src/redux/actions/settings.ts
@@ -64,3 +64,4 @@ export const settingsSetWorkflowAlertsOnErrorRecipients = createAction<string>(
 );
 export const settingsSetLayoutDirection = createAction<string>('settings/set-layout-direction');
 export const settingsSetEdgeType = createAction<string>('settings/set-edge-type');
+export const settingsSetSnapToGrid = createAction<boolean>('settings/set-snap-to-grid');
